requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
yfinance==0.2.18
pandas==2.1.4
numpy==1.25.2
xgboost==2.0.3
scikit-learn==1.3.2
ta==0.10.2
google-generativeai==0.3.2
langchain-core==0.1.7
python-dotenv==1.0.0
